requirements.txt
# SmartSpace UPY - Django Application
# Production Dependencies

# Django Core
Django>=5.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
gunicorn>=21.2.0
whitenoise>=6.6.0

# Django Extensions
django-unfold>=0.20.0
Pillow>=10.1.0

# Email
resend>=2.0.0

# Scheduler
apscheduler>=3.10.0

# AI Chatbot
google-generativeai>=0.3.0
